[S3] Add retry logic to delete-bucket

+++ src/aws_toolbox/commands/s3/delete_bucket.py
@@ -1,7 +1,9 @@
 import re
+import time
 
 import boto3
 import click
+from botocore.exceptions import ClientError
 
 from aws_toolbox.utils import logutils
 
@@ -45,17 +47,28 @@ def delete_bucket(ctx, name, dryrun):
         log.info(f"Emptying Bucket {bucket_name} with creation date {creation_date}")
 
         try:
-            object_versions = s3_client.list_object_versions(Bucket=bucket_name)
+            object_versions = s3_client.list_object_versions(Bucket=bucket_name).get("Versions", [])
             objects_to_delete = [
-                {"Key": obj["Key"], "VersionId": obj["VersionId"]} for obj in object_versions["Versions"]
+                {"Key": obj["Key"], "VersionId": obj["VersionId"]} for obj in object_versions
             ]
-            s3_client.delete_objects(Bucket=bucket_name, Delete={"Objects": objects_to_delete, "Quiet": False})
+            if objects_to_delete:
+                s3_client.delete_objects(Bucket=bucket_name, Delete={"Objects": objects_to_delete, "Quiet": False})
         except Exception as e:
             log.error(f"Cannot empty Bucket {bucket_name}: {e}")
 
         log.info(f"Deleting Bucket {bucket_name} with creation date {creation_date}")
 
-        try:
-            s3_client.delete_bucket(Bucket=bucket_name)
-        except Exception as e:
-            log.error(f"Cannot delete Bucket {bucket_name}: {e}")
+        max_attempts = 3
+        for attempt in range(1, max_attempts+1):
+            try:
+                s3_client.delete_bucket(Bucket=bucket_name)
+                log.info(f"Bucket {bucket_name} deleted")
+            except ClientError as e:
+                if e.response["Error"]["Code"] == "BucketNotEmpty":
+                    log.error(f"Cannot delete Bucket {bucket_name} (attempt {attempt}/{max_attempts}): {e}")
+                    time.sleep(1)
+                else:
+                    break
+            except Exception as e:
+                log.error(f"Cannot delete Bucket {bucket_name} (attempt {attempt}/{max_attempts}): {e}")
+                break
